ASoC: soc-pcm: fix a codec fixup issue in TDM case
On one intel platform, tx_mask is different with rx_mask. If
playback is started after capture, there is no sound produced.
It is caused by params fixup which changes the channel number
of playback based on rx_mask in hw_params for codec. And there
is no such issue if playback is started before capture.

According to the author of tx|rx_mask, tx_mask is for playback
and rx_mask is for capture. And stream direction would be checked
at other references of tx|rx_mask in ASoC, so here should be an error.
This issue could be rarely reproduecd because most drivers in kernel
set the same channel number to tx|rx_mask or rx_mask is zero.

Test on Intel platforms

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>