ASoC: intel: skl: fix a simultaneous playback & capture issue on hda platform
If playback and capture are working simultaneously and
capture is stopped, there is no sound produced by playback.
The register LOSIDV is only for playback to map output stream
for specific link but now it is set by capture.

This patch add check for stream direction before setting LOSIDV.

Test on WHL.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>