ASoC: SOF: Intel: hda: add SoundWire IP support
The helper calls the SoundWire initializations and sets the resource
structure defined as an interface with the platform driver. The
resource deals with PCI base address, interrupts and the callback for
the stream configuration.

hardware-related info: the interrupts are gated/ungated at the
top-level with the ADSPIPC2.SNDW field. The Core0 needs to be powered
before the SoundWire IP is initialized. We only have one context, but
depending on the hardware capabilities and BIOS settings may enable
multiple SoundWire links.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>