[HACK] soundwire: cadence_master: add debugsfs for MCP_COMMAND
The command FIFO can be accessed by the host, add entry to push
commands.  This is intended for hardware bring-up only and test the
command/control protocol, before drivers are fully functional.

The response support will be implemented with a read-only entry in a
follow-up patch.

This can be used in combination with error injection in cnds-cmd-ctrl
to see e.g. if the commands are taken into account (they shouldn't!)

The main use case are changing the frame shape to see if Slave devices
remain attached or reset them with in-band controls.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>