soundwire: cadence_master: provide plan B solution for state detection on errors
The Cadence IP provide a state detection mechanism, and will generate
an interrupt when the state of an external device changes. This is
quite efficient and helps avoid periodic polling.

In case an error happens while updating Slave states in the interrupt
handler/thread, then the device may perfectly be parked forever in the
same state and never be rechecked again.

This patch suggests an alternate solution triggered when errors
happen. In that case, during the next interrupt handled by the Cadence
IP, the driver will force a re-evaluation of the device states and use
the state reported by PING frames instead of the interrupt status.

If there are no interrupts generated, this patch will not provide a
satisfactory solution, and scheduling dummy commands on the bus will
be required to force the state re-evaluation.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>