[RFC] [TEST NEEDED] ASoC: SOF: Intel: hda: don't disable IPC interrupts in MSI mode
This patch is a POC to test Guennadi's comments in PR 1042.

The solution suggested in this patch keeps the existing code flow for
the PCI legacy mode, and avoids losing interrupts for the MSI mode.

MSI interrupt handling is currently conceptually broken. By
re-enabling the interrupt at the end of the thread, we may lost some
interrupts. We could modify the flow to re-enable the interrupt before
reading status registers, however this would be sub-optimial if we
used the PCI legacy mode since it would cause two calls to both the
IRQ handler and thread.

Another simpler solution would be to keep the interrupt enabled. This
would not be a problem for MSI, but would create an interrupt rain in
legacy mode. Having a conditional disabling/re-enabling of the
interrupt in the legacy mode solves that problem.

Let's test and see how this works.

Suggested-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>