soundwire: intel: add synchronize_irq() to avoid MSI timeouts
In MSI mode we keep having random timeouts, and with more links
enabled the probability of timeouts is increased significantly,
hinting at missing interrupts and handling of the MSI edge.

Adding a call to synchronize_irq() before enabling interrupts does
seem to make a difference and timeouts seem to be gone.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>