ASoC: SOF: Intel: add SoundWire configuration interface
Now that the SoundWire core supports the multi-step initialization,
call the relevant APIs.

The actual hardware enablement can be done in two places, ideally we'd
want to startup the IP as soon as possible but we need to work-out
additional hardware dependencies in corner cases. For the time being
the default is to startup the IP after the firmware is downloaded.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>