ASoC: SOF: Intel: hda: Disable L1 support during capture
A known issue in Host DMA results in xruns during
capture when the stream is released from pause.
The suggested workaround to address the issue is to
disable L1 support when there is an active capture
stream.

This issue does not affect all HDaudio platforms and seems to be
partly correlated to BIOS settings, e.g. the xruns do not happen on a
GeminiLake Chromebook using Coreboot but does show on the same
hardware with legacy BIOS w/ Ubuntu. Likewise, such underruns were
observed on IceLake SoundWire-based platforms using different BIOS
versions.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>