soundwire: bus: add wait_for_completion to prevent access to unenumerated Slaves
on resume, regmap immediately generates accesses to the Slaves, which
can race with enumeration. Add an explicit wait_for_completion()

FIXME: we should be able to do this unconditionally. For some reason
the initial probe results in timeouts.

FIXME: this works with RT711 but not with RT1308, possible paging
issue?

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>