ASoC: codecs: pcm512x: support GPIO-controlled clocks
Some PCM512x boards, such as HifiBerry DAC+ PRO, provide local
oscillators controlled by GPIOs.

Cross-dependencies make it hard to split the clock management from the
codec driver. The codec driver exposes a GPIOs, but the clock drivers
needs access to the codec regmap to verify if the clocks are present
and the codec locked. The codec driver also called devm_clk_get() in
its probe.

The suggested solution is to keep the codec and clock management in
the same module, with a conditional compilation. In its probe, the
codec driver creates the clock parts and passes the regmap needed by
the clock handling.

Note that this solution is not aligned with what is used in the RPI
tree, however the clock driver currently does absolutely nothing in
terms of hardware control, and all the handling is pushed to the
machine driver which manually accesses regmap and toggles GPIOs -
which beats the purpose of using the Common Clock Framework.

Tested on Up2 board with HifiBerry DAC+ and DAC+ PRO boards.

FIXME: the clocks are not detected for some reason

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>