soundwire: bus: register device on sdw_bus_master_add()
The existing code uses a pointer to device, which leads to problems
for sysfs support. Let's use a real device, registered on
sdw_bus_master_add() and unregistered on sdw_bus_master_delete().

This device will make it possible to use container_of() to go back to
the bus structure, which was really lacking for sysfs.

The Intel code is modified to avoid compilation errors but has to be
redone completely to remove platform devices.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>