ASoC: SOF: core: harden shutdown helper
When the probe is handled in a workqueue, we must use
cancel_work_sync() in the shutdown helper to avoid possible race
conditions.

Likewise, when a failure occurs in a probe workqueue, we should not
try to access hardware registers and create more problems. The
probe_continue() function already goes back to a blank slate on errors
so there's nothing to do for a shutdown operation.

Fixes: daff7f1478e12 ("ASoC: SOF: add snd_sof_device_shutdown() helper for shutdown")
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>