soundwire: stream: update prototypes before split of add into alloc/config
In order to become compatible with ALSA/ASoC expectations that the
hw_params() can be called multiple times, the SoundWire stream API
needs to be updated so that the memory allocation and configurations
are split. The memory alloc/remove steps will be called in the
.startup and .shutdown callbacks respectively, while the configuration
can take place in the .hw_params or .prepare callbacks.

This patch adds the new alloc/config APIs, and changes the return type
of the remove to void. No functional change otherwise.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>