[WIP] soundwire: stream: implement add with alloc/config routines
Before splitting the code in two, this patch exposes two new functions
to alloc/configure the stream on the master and peripheral sides. This
is a temporary step to make sure there are no regressions. In a future
steps, the master and peripheral drivers will directly use the
alloc/config functions.

The functions are re-grouped by master_rt, slave_rt, port_rt
types. The _rt functions are renamed to consistently use the
type-action convention.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>