ASoC: SOF: Intel: hda-dai: split dailink and dai operations
First split the dailink and dai operations in separate functions, that
will still be called from the dai callbacks.

In an ideal world, these dailink routines would be moved and invoked
directly from the dailink callbacks. This would be quite invasive and
impact multiple machine drivers, so for now we keep the split internal
to the SOF driver. The main intent here is to split 'link management'
from 'SOF IPC'.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>