ASoC: SOF: ipc3-topology: add quirk for Intel SSP mclk_id
The mclk_id is specified in the topology. 95% of devices typically use
the MCLK_ID 0. In rare cases, some Chromebooks use MCLK_ID 1.

For the ES8336-based platforms, the same ratio applies but we have no
information from NHLT on what the mclk_id is. It might be possible to
fetch an NHLT blob and reverse-engineer the registers, but that's a
lot of work for limited benefits.

This patch suggests a kernel parameter instead, where users can try to
override the default mclk_id, and it saves us the hassle of
maintaining yet another permutation for these ES8336 platforms.

options snd_sof sof_ssp_mclk_id=1

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>